requirements.txt
pytest
selenium
pyyaml
certifi
bs4
html5lib
playsound
requests